@@ -1,6 +1,6 @@
 {
   "name": "ts-jest",
-  "version": "29.1.5",
+  "version": "29.2.3",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "bin": {
@@ -11,7 +11,7 @@
     "prebuild": "rimraf dist coverage *.tgz",
     "build": "tsc -p tsconfig.build.json",
     "postbuild": "node scripts/post-build.js",
-    "pretest": "tsc -p tsconfig.spec.json --noEmit && node scripts/prepare-test.js",
+    "pretest": "tsc -p tsconfig.spec.json --noEmit && node scripts/create-bundle.js",
     "test": "jest",
     "test-examples": "node scripts/test-examples.js",
     "lint": "eslint --ext .js,.ts .",
@@ -52,6 +52,7 @@
   "homepage": "https://kulshekhar.github.io/ts-jest",
   "dependencies": {
     "bs-logger": "0.x",
+    "ejs": "^3.1.10",
     "fast-json-stable-stringify": "2.x",
     "jest-util": "^29.0.0",
     "json5": "^2.2.3",
@@ -93,25 +94,26 @@
     }
   },
   "devDependencies": {
-    "@commitlint/cli": "^19.3.0",
-    "@commitlint/config-angular": "^19.3.0",
+    "@commitlint/cli": "18.6.1",
+    "@commitlint/config-angular": "18.6.1",
     "@jest/transform": "^29.7.0",
     "@jest/types": "^29.6.3",
     "@types/babel__core": "7.20.5",
     "@types/cross-spawn": "latest",
+    "@types/ejs": "^3.1.5",
     "@types/fs-extra": "latest",
     "@types/js-yaml": "latest",
     "@types/lodash.camelcase": "4.3.9",
     "@types/lodash.memoize": "4.1.9",
     "@types/lodash.set": "4.3.9",
-    "@types/micromatch": "4.0.7",
-    "@types/node": "20.14.2",
+    "@types/micromatch": "4.0.9",
+    "@types/node": "20.14.11",
     "@types/react": "18.x",
     "@types/semver": "latest",
     "@types/yargs": "latest",
     "@types/yargs-parser": "21.0.3",
-    "@typescript-eslint/eslint-plugin": "^7.13.0",
-    "@typescript-eslint/parser": "^7.13.0",
+    "@typescript-eslint/eslint-plugin": "^7.16.1",
+    "@typescript-eslint/parser": "^7.16.1",
     "babel-jest": "^29.7.0",
     "conventional-changelog-cli": "^5.0.0",
     "cross-spawn": "latest",
@@ -120,9 +122,9 @@
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-import": "^2.29.1",
     "eslint-plugin-jest": "^28.6.0",
-    "eslint-plugin-jsdoc": "^48.2.12",
+    "eslint-plugin-jsdoc": "^48.7.0",
     "eslint-plugin-prefer-arrow": "^1.2.3",
-    "eslint-plugin-prettier": "^5.1.3",
+    "eslint-plugin-prettier": "^4.2.1",
     "execa": "5.1.1",
     "fs-extra": "11.2.0",
     "glob": "^10.2.6",
@@ -133,11 +135,8 @@
     "js-yaml": "latest",
     "json-schema-to-typescript": "^13.1.2",
     "lint-staged": "latest",
-    "lodash.camelcase": "^4.3.0",
-    "lodash.set": "^4.3.2",
-    "node-fetch": "^3.3.2",
-    "prettier": "^3.3.2",
-    "typescript": "~5.1.3"
+    "prettier": "^2.8.8",
+    "typescript": "~5.5.3"
   },
   "lint-staged": {
     "*.{ts,tsx,js,jsx}": [
